fix  bugs: 1. set nullptr when press NCL::MouseButtons::Middle  (tutorialgame) 2. pass same type of bullet triggers (BulletContactListener::onTrigger)
freelist is supported in network actually

+++ CSC8503/TutorialGame.cpp
@@ -217,9 +217,12 @@ void TutorialGame::UpdateKeys() {
 
 	if (Window::GetMouse()->ButtonPressed(NCL::MouseButtons::Middle)) {
 		if (G1.selectionObject) {
-			G1.selectionObject->SetActive(false);
-			G1.DeleteObject(G1.selectionObject);
-			//G1.selectionObject = nullptr;
+			if(true)
+				G1.selectionObject->SetActive(false);
+			else {
+				G1.DeleteObject(G1.selectionObject);
+				G1.selectionObject = nullptr;
+			}
 		}
 	}
